moving article_elements functions to separate file
creating article_elements.py, fixing import statements
